Separation of diagram and diagram control
